Move test to its own file
